Move functions for creating TarHeaders into tarHeader.ts
